More robust results.lib.php
There's a bug lurking in the DB ingest which can result in double-entries for a single UTC. This can affect the results page as not enough rows would get generated when displaying the first page and then catastrophic failure would occure when requesting further pages. Current design should be robust against this.